Some improvements to let institutional admins register new users.
refs #1362
