--- app/views/admins/_form.html.haml
@@ -1,14 +1,0 @@
-= simple_form_for @user, url: admin_create_user_path, html: { class: 'single-column' } do |f|
-
-  .modal-body
-    #user-basic-info
-      = f.input :email, required: true, autofocus: true
-      = f.input :_full_name, required: true
-      = f.input :username, required: true
-      - if current_user.superuser?
-        = f.input :institution_id, required: true
-      = f.input :password, required: true
-      = f.input :password_confirmation, required: true
-
-  .modal-footer
-    = f.button :wrapped, cancel_modal: true, value: button_name, class: "btn btn-primary"
