--- app/views/admins/new_user.html.haml
@@ -1,5 +1,0 @@
-#user-new
-  .modal-header
-    = modal_close_button
-    %h3= t('.title')
-  = render 'form', :button_name => t('_other.create')
