+++ app/views/users/new.html.haml
@@ -10,6 +10,8 @@
         = f.input :email, required: true, autofocus: true
         = f.input :_full_name, required: true
         = f.input :username, required: true
+        - if current_user.superuser?
+          = f.input :institution_id, required: true
         = f.input :password, required: true
         = f.input :password_confirmation, required: true
